extract movie classes and add reset functionality
